Make SpreadsheetMLReader an inner class
